+++ src/shared/Network/ListenSocketLinux.h
@@ -40,14 +40,14 @@ public:
         int ret = bind(m_socket, (const sockaddr*)&m_address, sizeof(m_address));
         if(ret != 0)
         {
-            printf("Bind unsuccessful on port %u.", Port);
+            printf("Bind unsuccessful on port %u.", (unsigned int)Port);
             return;
         }
 
         ret = listen(m_socket, 5);
         if(ret != 0) 
         {
-            printf("Unable to listen on port %u.", Port);
+            printf("Unable to listen on port %u.", (unsigned int)Port);
             return;
         }
         len = sizeof(sockaddr_in);
